+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- `ConditionBasedLoopControl` that is used on every tick of the loop to stop it when condition is met
 
 [Unreleased]: https://github.com/ecomdev/react-event-loop-test-utils/compare/4b825dc642cb6eb9a060e54bf8d69288fbee4904...HEAD